extensions/data-plane/data-plane-azure-storage/src/main/java/org/eclipse/edc/connector/dataplane/azure/storage/DestinationBlobName.java
/*
 *  Copyright (c) 2024 Microsoft Corporation
 *
 *  This program and the accompanying materials are made available under the
 *  terms of the Apache License, Version 2.0 which is available at
 *  https://www.apache.org/licenses/LICENSE-2.0
 *
 *  SPDX-License-Identifier: Apache-2.0
 *
 *  Contributors:
 *       Marco Primo
 *
 */

package org.eclipse.edc.connector.dataplane.azure.storage;

import org.eclipse.edc.util.string.StringUtils;

/**
 * Utility class responsible for determining the name under which a file will be saved.
 */
public class DestinationBlobName {

    private final String folderName;
    private final String blobName;

    public DestinationBlobName(String blobName, String folderName) {

        this.blobName = blobName;
        this.folderName = folderName;
    }

    /**
     * Resolves the name under which a resource should be saved based on its part name and size.
     *
     * @param partName The name of the resource part.
     * @param partsSize The size of the resource part.
     * @return A String representing the resolved name for the resource.
     */
    public String resolve(String partName, int partsSize) {

        var name = (partsSize == 1 && !StringUtils.isNullOrEmpty(this.blobName)) ? this.blobName : partName;
        if (!StringUtils.isNullOrEmpty(this.folderName)) {
            name = this.folderName.endsWith("/") ? this.folderName + name : this.folderName + "/" + name;
        }
        return name;
    }

}

